feat(post): post popup dialog, advanced share post modal sheet
Added a lot of minor and major improvements in performance and code quality.

Added feature to share post with multiple people at a time. Need to add support of a display of the message, attached to the shared post as a message.

Added feature to preview a post in user profile view on long press. A modal with post details pops up. Depending on the last reletive position to the buttons such as like, comment, share or options, the appropriate function will be executed, after a smooth dialog close animation.

Replaced almost all existing blocs from `Bloc` to `HydratedBloc`. Removed direct access to stream from blocs, instead created events that creates subscriptions to the appropriate stream(made seperation of concerns).