feat(post): edit post + minor chore changes
A lot of minor fixes, improvements and alters. Replaced `Colors` to `AppColors` in codebase. Replace all static `EdgeInsets` padding values to `AppSpacing`.

Added feature that allows to edit a post.

Now, whenever a new post is created/deleted/updated it will be respectively displayed/updated/deleted in `FeedPage` as well as in `ReelsPage`.