fix: added explicit `onUserTap` callback
There was an issue, tapping on username of first user that liked the post and is in following of current user the navigator should push that user profile, however it always pushed profile by `BlockAction.authorId`. So I added an explicit `onUserTap` callback that fixed that problem and now it correctly navigates to that user profile screen.