fix: fixed an issue with range error when send first message
The issue was caused by scrolling to the bottom of the chat page using `itemScrollController` before there was first item to actually scroll to. Due to that, when sending first message there was a range error, saying there was no allowed index. This was fixed by adding 50 milliseconds delay before scrolling to the bottom of the chat page.

Fixed range error that was caused on `FloatingDateSeparator`. The issue appeared when there was not enough messages to actually scroll chat page. In that case after user deleted a message when the function `getTopElementIndex` was executed it returned the index equal to the length of all messages, causing the range error. It was fixed by checking if index equals to the messages length then calculate the index like: `messages.length - index` instead of `messages.length - 1 - index` if reversed.