chore: check if story was visible before playing
When `visibleBounds` were not empty when `StoryView` was opened the `StoryController` was set to `play`. It is ok to happen when pushing screen on top of `StoriesPage` and pop it back to play stories back, but it also played story when it was loading the data just becaue `visibleBounds` were not empty. With explicit `wasVisible` flag we can check if the story was not seen to accordingly play story when `visibleBounds` are not empty.