chore: added localization for `storied_editor`
Added localization for `stories_editor`.

Fixed an unexpected behaviour when the `ChatPage` was rebuilding multiple times the keyboard appeared.

A little bit tweaked the view of `ReelView`.

Now the preffered orientation is explicitly set to only `DeviceOrientation.portraitUp` and `DeviceOrientation.portraitDown`, so that user no more can landscape orientation, because the application is not intended to be used in the orientation and it doesn't have an adaptive design to support this feature, so it is disabled now.