Create LICENSE
The code for this project lays under MIT-License and is in open source, open for modifications and contributions. However, important note: if you want to copy and modify the source code of my project and publish it somewhere, you MUST link me as an original author of the code AS IS Emil Zulufov. Without specifying it, you may potentially have some problems. Thanks for understanding!

Signed-off-by: Emil Zulufov (ezIT) <emilzulufov@gmail.com>